The long experiment constraint - improved, it now penalises single allocations of long constraints.
Each of the beamlines available should be used at least once - names
hard-coded in several new hard constraint has been added. There should
be a neat way of doing it!!! Investigate.

The sum of preferences has bee doubled so that the duplicated weights of
the long experiments will never result in positive weights - there
should be a nicer way of avoiding this. Temporary solution??